% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_year.R
\name{cor_year}
\alias{cor_year}
\title{Correlation of Basketball Statistics}
\usage{
cor_year(year = 1983, corrgram = FALSE)
}
\arguments{
\item{year}{Input a year between 1950 and 2017. Defaults to 2017 and 
generates a random year if input is out of range.}

\item{corrgram}{If you want a corrgram (requires "corrgram" package) returned,
set parameter to TRUE. Defaults to FALSE.}
}
\description{
This function returns the oldest basketball player in a year. If there
is more than one player with the same age, it returns one random player
of the oldest age.
}
\examples{
cor_year(2004)
cor_year(2005, corrgram = TRUE)
}
\keyword{age}
\keyword{nba,}
\keyword{statistics,}
