% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_year.R
\name{fact_year}
\alias{fact_year}
\title{Random NBA Fact of the Year}
\usage{
fact_year(year = 1983)
}
\arguments{
\item{year}{Input a year between 1950 and 2017. Defaults to 2017 and 
generates a random year if input is out of range.}
}
\description{
This function returns a random NBA fact for the input year.
}
\examples{
fact_year(1989)
}
\keyword{nba,}
\keyword{statistics}
