% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_year.R
\name{oldest_year}
\alias{oldest_year}
\title{Get the Oldest Basketball Player by Year}
\usage{
oldest_year(year = 1983)
}
\arguments{
\item{year}{Input a year between 1950 and 2017. Defaults to 2017 and 
generates a random year if input is out of range.}
}
\description{
This function returns the oldest basketball player in a year. If there
is more than one player with the same age, it returns one random player
of the oldest age.
}
\examples{
oldest_year(2002)
}
\keyword{age}
\keyword{nba,}
\keyword{statistics,}
